+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/tables/cells/DiscoveredAssetDataUseCell.tsx
@@ -25,12 +25,14 @@ const DiscoveredAssetDataUseCell = ({
 
   const { getDataUseDisplayName } = useTaxonomies();
 
+  const currentDataUses =
+    asset.user_assigned_data_uses || asset.data_uses || [];
+
   const handleAddDataUse = async (newDataUse: string) => {
-    const existingUses = asset.user_assigned_data_uses || asset.data_uses || [];
     const result = await updateAssetsDataUseMutation({
       monitorId: asset.monitor_config_id!,
       urnList: [asset.urn],
-      dataUses: [...existingUses, newDataUse],
+      dataUses: [...currentDataUses, newDataUse],
     });
     if (isErrorResult(result)) {
       errorAlert(getErrorMessage(result.error));
@@ -44,11 +46,10 @@ const DiscoveredAssetDataUseCell = ({
   };
 
   const handleDeleteDataUse = async (useToDelete: string) => {
-    const existingUses = asset.user_assigned_data_uses || asset.data_uses || [];
     const result = await updateAssetsDataUseMutation({
       monitorId: asset.monitor_config_id!,
       urnList: [asset.urn],
-      dataUses: existingUses.filter((use) => use !== useToDelete),
+      dataUses: currentDataUses.filter((use) => use !== useToDelete),
     });
     if (isErrorResult(result)) {
       errorAlert(getErrorMessage(result.error));